deleted interface, service class adjusted
