+++ src/main/java/guestRegister/controllers/HomeController.java
@@ -1,8 +1,6 @@
 package guestRegister.controllers;
 
 
-import guestRegister.models.services.GuestService;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
